AbstractAlphanumericForm to check fields from form
AbstractAlphanumericForm was checking the modified fields
using the ones from the file model. This raised an error
if not all of them have a field in the form, so it has been
changed to get the fields from the form, look for them in
the file and if they exist, check them as it was already doing.