Support for key-value strategy in combobox.
Sometimes, comboboxes need to show some string value but when saving it, this value needs to be converted to other (an ID key for example). That is very common if you use a database for keeping the domain values in a separate table of the database. You want to show the string values to the user, but saving them as ID. The key-value strategy is a good solution for that.

See: http://stackoverflow.com/questions/2095140/need-flexible-java-key-value-collection-class-for-jcombobox