Component's name is no longer uppercased:
when working with a layer containing thousand of registers (I'm playing with 40.000 now), I've experienced a meaningful gain in performance without uppercasing the strings. This change will obligue us to be aware of the casing in the abeille form and the field in the database, having to be both equals. I recognised this as a nuisance, but I think is worth the change in order to make the forms usable for big amount of registers.