ntforms behaves as a library so, it should not contain either UI or i18n resources inside it.
If we place i18n resource in the library, in order to retrieve them using gvSIG conventions and tools (for example: PluginServices.getText() ) we should deploy ntforms as a extension within gvSIG workspace. In order to avoid this and make ntforms behaves as a library, the UI and i18n resources are dropped and they should live in other places (navtable, the extension inheriting from ntforms, etc).