Use the class as identifier for the logger
Use the name of the class as identifier for the logger is a good
programming practice. With this change the use of the child class
as identifier is forced, and less configuration on the child is need.
The logger field is defined as protected to allow be used by childs