Subforms clone the controller & do CUD with it
In order to have multiple views operating with the
same source, we had to provide support for multiple
controllers. Now the subform clones and uses it for
the creating, deleting and updating rows. After
those changes, the model, which is the one that
holds the listeners, is warned that the data source
has changed.