Fixes on the persistance of size and position of forms
The list of windows properties should be read from the persistance
file each time we are going to use it, because the the persistance
file can be written from other opened forms or subforms. So instead
of having it as a field is handled as a local variable that is
reseted always.