Heavy refactoring of table model classes.
Now we have an specific package for table models
related classes. This includes a new TableModelBase,
the TableModelAlphanumeric, which now extends the
previous one, a new TableModelVectorial, the Not-
EditableTableModel and the TableModelFactory.
The table models methods have also been slightly
refactored in order to fit in the new hierarchy
while avoiding redundant code, as well as for
accepting row filters.