TableHandlers management in AbstractForm.
Now that all the TableHandlers have the same interface,
we can manage them in a generic way. Subclasses should
add them with the addTableHandler method, and override
getPrimaryKeyValue which must return the current value
for that id. Then by iterating the list all handlers
are filled with that value, and the listeners are set
and removed.
Due to these changes, the positionChanged event is now
processed in order to call fillSpecificValues.