TocTableManager can now retrieve tables by models.
TocTableManager used the opened windows for searching
for the tables, which wouldn't work if you simply closed
the window but left the table open in the project defi-
nition. That function is kept just in case it's needed
somewhere, but a new one has been added which adds the
desired functionality, retrieving the project and its
opened project models for searching among them for the
desired table.
Also changed all the calls in the new NavTableForms'
handlers structure to the new method.