Make sure the IController has the default cb value
While filling comboboxes with domainvalues sometimes
the IController didn't have the default (initial) cb
value inside it when opening the form. This happened
when the record was new (or somehow had a previous
null value in that field) and the combobox didn't
accept the null or void value. In that case the
iController started with a void value (the one obtai-
ned from the blank record) whereas the combobox displays
a different one as default. Therefore saving without
clicking onto the combobox would have saved that null
value in that field.